feat(domain): refactor Invoice aggregate with DDD
enforce encapsulation (remove setters, protect _items)
model lifecycle via InvoiceStatus (Draft → Issued → Paid / Cancelled)
implement domain behaviors (AddItem, Issue, Pay, Cancel) with validation
introduce domain events for key transitions
standardize structured Error handling (Code + Message)

+++ Domain/Invoices/Events/InvoiceCreatedEvent.cs
@@ -10,6 +10,6 @@ namespace Domain.Invoices.Events
 {
     public record InvoiceCreatedEvent(
         InvoiceId id,
-        double Price
+        decimal Price
         ):DomainEvent;
 }
